the_eye/pyproject.toml
[tool.poetry]
name = "the_eye"
version = "0.1.0"
description = ""
authors = ["Francisco Alface <falface@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.2"
python-decouple = "^3.5"
celery = "^5.2.1"
gunicorn = "^20.1.0"
redis = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"
pytest-cov = "^3.0.0"
black = "^21.11b1"
isort = "^5.10.1"
bandit = "^1.7.1"
safety = "^1.10.3"
pytest-django = "^4.4.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
